requirements/base.txt
django
Pillow
celery
redis
pandas
djangorestframework
slackclient
